iio: adrv9002: add checks for ORx
ORx can only be enabled if RX is not enabled and TX is enabled. On top
of that, after enabling it, we won't allow any operation that can
trigger an ensm transition in RX/TX of the same channel as the enabled
ORx. Because allowing it, has the non explicit side effect of
potentially break the ORx capture...

Also note that with this, we need to enable the ORx tracking cal
(if desired) before enabling the capture.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>
(cherry picked from commit 316fe0292362d450250d22080ee794d4e3f55f95)