iio: adrv9002: update driver for new sdk API
There are two mains changes with the update:

1) It is now clear that the RX gain table type cannot change at runtime.
This means that the ABI for the digital interface gain was wrong and
so, it was fixed (which resulted in a different ABI). On top of this,
for now, there's no way to get the table type from the profiles which
means we are defaulting to correction mode (which was the one used
before anyways). This info might become available in the profile
structure in next SDK releases so that I'm not rushing in having a
different, non ideal way of selecting the table type.

2) It is now possible to select Single Data rate ports for LVDS.
However, this is not yet supported in hdl so that we won't allow such
profiles to be loaded...

Signed-off-by: Nuno Sá <nuno.sa@analog.com>
(cherry picked from commit 63af6a270bb2b361cc5e0c152794a77614b34055)