iio: adrv9002: adapt to the new API
There are two mains things that had to be changed:

1) Gain setup by pin in Frequency hopping can now be done by channel.
Hence the devicetree parsing and the table load attr had to be changed
accordingly. Ditto for the debugfs interface.

2) The digital interface gain has now RSSI parameters. So, when changing
the control mode we need to get the complete structure and only
change the mode. Otherwise we would get an error from the API since
these values cannot be 0.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>