configs: Fix ethernet connection issues for MitySOM A10s Devkit
Ethernet drivers require the Altera System Manager drivers which
are not enabled by default from 5.x kernels. The current socfpga_adi_defconfig
is based on the 4.14 kernels which does not need this config to start
the syscon drivers. Hence, adding the MFD_ALTERA_SYSMGR config in the
defconfig to enable the SOCFPGA System Manager drivers in kernel
and enable ethernet properly by fixing the following errors.

  socfpga-dwmac ff802000.ethernet: No sysmgr-syscon node found
  socfpga-dwmac ff802000.ethernet: Unable to parse OF data

Signed-off-by: d-samal <dibyajyoti.samal@einfochips.com>