adc: navassa: Add patch file for MitySOM bootup
MitySOM A10s seems to have tuning bootup errors even though
there are no data acquisition errors in ADRV9002.
Apply this patch file to set the delay values to
driver defaults since the same test pattern is used for
caliberations. This not a long term patch, needs to be fixed
and reverted once HDL does not need this patch.

Signed-off-by: d-samal <dibyajyoti.samal@einfochips.com>