iio: adc: adrv9002: add macro to wrap api calls
The main motivation for this change was to avoid patterns like:

```
if (ret)
	return adrv9002_dev_err(phy);

return 0
```

or

```
if (ret)
	ret = adrv9002_dev_err(phy)
```

The reason why the above is needed is to translate api error codes to
linux codes. Hence, just wrap that logic in an helper macro which allow
us to actually save some significant amount of lines of code.

While at it, two mains things were also done:

1) Conformed to 100 col limit (when the driver was first added, it was
still doing 80)

2) Use ternary operator for patterns like 'return ret ? ret : foo'
instead of:

```
if (ret)
	return ret;

return foo
```

There were a mix of style regarding this, so let's try to be more
consistent and just choose a style (I might still have missed some
places).

Signed-off-by: Nuno Sá <nuno.sa@analog.com>