iio: adc: adrv9002: constify where possible
There were a lot of places where the objects could be declared as const.
Having them const is obviously a good practise as, among other things,
allows the compilers to better optimize the code.

Note this is a work in progress so there might be other places that
could use this improvement (notably adrv9002_conv.c and
adrv9002_debugfs.c).

While at it, changed the pattern 'struct iio_chan_spec const *chan' to
'const struct iio_chan_spec *chan' so it's consistent. On top of this,
turned adrv9002_axi_digital_tune_verbose() static.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>